Renormalize line endings
Ran `git add --renormalize .`

A few files were stored with CRLF line endings, when according
to .gitattributes they will be stored with LF line endings next
time a change is committed.

This was making it so any edit to these files would result in the
_entire_ file being in the diff.

Renormalize these files so there are no more problems and line
endings work correctly according to .gitattributes.