Fix world wrap not wrapping properly
World wrap was not wrapping left side onto right
https://github.com/Creatures-Developer-Network/C2toDS-Refresh/issues/7#issuecomment-1073063784

This update creates a wrap area equal to the size of the window. The wrap size changes as the game window is resized.

Initial construction of the wrap is delayed as eager construction caused right side panels to be placed at 0.00.

References #7