removed dependence on otherwise unused sprite file
The only thing pointing to "moe_C2toDS_fish" was the larvae stage of the jellyfish, so I've replaced it with a new white tinted version of the goldfish egg at the end of the jellyfish sprite file. Previously it was indistinguishable from the goldfish egg.

So this is better, but still a placeholder really.